@@ -1,8 +1,7 @@
-/**
- * A grid file.
- */
+// @flow
+import type {Grid} from './types'
 
-export function create (data) {
+export function create (data: ArrayBuffer): Grid {
   const array = new Int32Array(data, 4 * 5)
   const header = new Int32Array(data)
 
