@@ -1,14 +1,18 @@
-/** perform math on web mercator grids, http://wiki.openstreetmap.org/wiki/Slippy_map_tilenames#Mathematics */
+// @flow
+/**
+ * Perform math on web mercator grids
+ * http://wiki.openstreetmap.org/wiki/Slippy_map_tilenames#Mathematics
+ */
 
 /** convert longitude to pixel value */
-export function lonToPixel (lon, zoom) {
+export function lonToPixel (lon: number, zoom: number): number {
   // factor of 256 is to get a pixel value not a tile number
   // return int, 32 bit int sufficient for pixels up to zoom 22
   return ((lon + 180) / 360 * Math.pow(2, zoom) * 256) | 0
 }
 
 /** convert latitude to pixel value */
-export function latToPixel (lat, zoom) {
+export function latToPixel (lat: number, zoom: number): number {
   const invCos = 1 / Math.cos(radians(lat))
   const tan = Math.tan(radians(lat))
   const ln = Math.log(tan + invCos)
@@ -17,22 +21,22 @@ export function latToPixel (lat, zoom) {
 }
 
 /** convert pixel value to longitude */
-export function pixelToLon (x, zoom) {
+export function pixelToLon (x: number, zoom: number): number {
   return x / (Math.pow(2, zoom) * 256) * 360 - 180
 }
 
 /** convert pixel value to latitude */
-export function pixelToLat (y, zoom) {
+export function pixelToLat (y: number, zoom: number): number {
   const tile = y / 256
   return degrees(Math.atan(Math.sinh(Math.PI - tile * Math.PI * 2 / Math.pow(2, zoom))))
 }
 
 /** convert radians to degrees */
-export function degrees (rad) {
+export function degrees (rad: number): number {
   return rad * 180 / Math.PI
 }
 
 /** convert degrees to radians */
-export function radians (deg) {
+export function radians (deg: number): number {
   return deg * Math.PI / 180
 }
