@@ -1,18 +1,26 @@
+// @flow
+import slice from 'lodash/slice'
+
+import {ITERATION_WIDTH, MAX_PATH_SEGMENTS} from './constants'
+
+import type {Origin, PathDescriptor, Point, StopPatternStops} from './types'
+
 /**
  * Represents the data for a single origin, including a per-transit-stop index.
  * @author mattwigway
- */
-
-/**
  * @param {ArrayBuffer} data
- * @param {Object} point
- * @param {Number} point.x
- * @param {Number} point.y
+ * @param {Point} point
+ * @returns {Origin}
  */
-
-export const ITERATION_WIDTH = 4
-
-export function create (data, point) {
+export function create ({
+  data,
+  log = console.log.bind(console),
+  point
+}: {
+  data: Int32Array,
+  log: Function,
+  point: Point
+}): Origin {
   const origin = {}
 
   origin.data = data
@@ -25,6 +33,7 @@ export function create (data, point) {
   origin.radius = data[offset++]
   const nPixels = Math.pow(origin.radius * 2 + 1, 2)
 
+  log(`origin radius: ${origin.radius} nPixels: ${nPixels}`)
   for (let pixel = 0, nonTransitTime = 0; pixel < nPixels; pixel++) {
     nonTransitTime += data[offset] // don't increment offset here, will be incremented on write
     data[offset++] = nonTransitTime // increment offset here
@@ -35,6 +44,7 @@ export function create (data, point) {
   const nMinutes = data[offset++]
   origin.nStops = nStops
   origin.nMinutes = nMinutes
+  log(`origin nStops: ${nStops} nMinutes: ${nMinutes}`)
 
   origin.index = new Int32Array(nStops)
 
@@ -76,22 +86,45 @@ export function create (data, point) {
   return origin
 }
 
-  /** get a non transit time in minutes, or 255 if you cannot walk to this pixel */
-export function getNonTransitTime (origin, {x, y}) {
-  let relx = x - origin.x
-  let rely = y - origin.y
+/**
+ * Get `StopPatternStop` set for a path descriptor
+ */
+export function getStopPatternStopSets ({
+ log = console.log.bind(console),
+ origin,
+ pathDescriptor
+}: {
+  log: Function,
+  origin: Origin,
+  pathDescriptor: PathDescriptor
+}): StopPatternStops {
+  const [stop, pathIdx] = pathDescriptor
+  if (stop === -1) return []
+  // offset to the first path, skipping times and nPaths
+  let offset = origin.index[stop] + origin.nMinutes * ITERATION_WIDTH + 1
 
-  if (Math.abs(relx) <= origin.radius && Math.abs(rely) <= origin.radius) {
-    // make them nonnegative so they can be used as offsets
-    relx += origin.radius
-    rely += origin.radius
-
-    // we can possibly walk to this pixel
-    // the index of the pixel, plus one because the radius is recorded at the start of the array
-    let timeSecs = origin.data[rely * (origin.radius * 2 + 1) + relx + 1]
-    let timeMins = timeSecs / 60 | 0
-    if (timeSecs !== -1 && timeMins < 255) return timeMins
+  // seek forward to correct path
+  let curIdx = 0
+  while (curIdx++ < pathIdx) {
+    const nSegments = origin.data[offset++]
+    offset += nSegments * 3
   }
 
-  return 255
+  const nSegments = origin.data[offset++]
+
+  if (nSegments > MAX_PATH_SEGMENTS) {
+    // bail if we have a ridiculous path (clearly a data/code error), and complain loudly
+    log(`Too many path segments (${nSegments} > ${MAX_PATH_SEGMENTS}) in path ${pathIdx} to stop ${stop} from origin ${origin.x}, ${origin.y}, returning no path. This implies a bug, please raise cain about it`)
+    return []
+  }
+
+  const path = []
+  for (let seg = 0; seg < nSegments; seg++) {
+    // not delta coded
+    // TODO is creating typed arrays which may be inefficient for small values
+    path.push(slice(origin.data, offset, offset + 3))
+    offset += 3
+  }
+
+  return path
 }
