@@ -1,10 +1,25 @@
+// @flow
+import {ITERATION_WIDTH} from './constants'
+import {getNonTransitTime} from './utils'
+
+import type {Origin, Query, StopTreeCache} from './types'
+
+type NextParameters = {
+  travelTimesForDest: Uint8Array,
+  walkTimesForDest: Uint8Array,
+  inVehicleTravelTimesForDest: Uint8Array,
+  waitTimesForDest: Uint8Array,
+  x: number,
+  y: number
+}
+
 /**
- * Perform propagation from an origin to destination cells, calling a callback to summarize the data.
+ * Perform propagation from an origin to destination cells, calling a callback
+ * to summarize the data. This seems like it's overly complicated, but I don't
+ * think there's an easier way to have this code used both to create surfaces,
+ * isochrones and spectrogram data, which are different summaries of the data.
  *
- * This seems like it's overly complicated, but I don't think there's an easier way to have this code
- * used both to create surfaces/isochrones and spectrogram data, which are different summaries of the data.
- *
- * The callback will be called with the following, once for each destination pixel
+ * The callback will be called once for each destination pixel with:
  *   travelTimesForDest: Travel times at each iteration
  *   walkTimesForDest: Walk times at each iteration
  *   inVehicleTravelTimesForDest: In vehicle travel times at each iteration
@@ -12,10 +27,17 @@
  *   x: X coordinate of grid for this destination
  *   y: Y coordinate of grid for this destination
  */
-
-import {getNonTransitTime, ITERATION_WIDTH} from './origin'
-
-export default function propagate ({ query, stopTreeCache, origin, callback }) {
+export default function propagate ({
+  next,
+  origin,
+  query,
+  stopTreeCache
+}: {
+  next: (NextParameters) => void,
+  origin: Origin,
+  query: Query,
+  stopTreeCache: StopTreeCache
+}) {
   const {nMinutes} = origin
   const travelTimesForDest = new Uint8Array(nMinutes) // the total travel time per iteration to reach a particular destination
   const waitTimesForDest = new Uint8Array(nMinutes) // wait time per iteration for particular destination
@@ -35,8 +57,10 @@ export default function propagate ({ query, stopTreeCache, origin, callback }) {
       travelTimesForDest.fill(nonTransitTime)
       waitTimesForDest.fill(255)
       inVehicleTravelTimesForDest.fill(255)
-      walkTimesForDest.fill(nonTransitTime) // when the origin is within walking distance and
-        // walking is the fastest way to reach the destination, _everything_ is walk time
+
+      // when the origin is within walking distance and walking is the fastest
+      // way to reach the destination, _everything_ is walk time
+      walkTimesForDest.fill(nonTransitTime)
 
       for (let stopIdx = 0; stopIdx < nStops; stopIdx++) {
         // read the stop ID
@@ -52,7 +76,8 @@ export default function propagate ({ query, stopTreeCache, origin, callback }) {
           if (travelTimeToStop !== -1) {
             const travelTimeToPixel = travelTimeToStop + time
 
-            // no need to check that travelTimeToPixel < 255 as travelTimesForDest[minute] is preinitialized to the nontransit time or 255
+            // no need to check that travelTimeToPixel < 255 as
+            // travelTimesForDest[minute] is preinitialized to the nontransit time or 255
             if (travelTimesForDest[minute] > travelTimeToPixel) {
               travelTimesForDest[minute] = travelTimeToPixel
               const inVehicle = inVehicleTravelTimesForDest[minute] = origin.data[offset + 1]
@@ -67,7 +92,7 @@ export default function propagate ({ query, stopTreeCache, origin, callback }) {
         }
       }
 
-      callback({
+      next({
         travelTimesForDest,
         walkTimesForDest,
         inVehicleTravelTimesForDest,
