@@ -1,3 +1,5 @@
+/* @flow */
+import type {StopTreeCache} from './types'
 /**
  * Cache trees from pixels to transit stops, including an index for where a particular pixel is located.
  * @author mattwigway
@@ -6,14 +8,18 @@
 /**
  * Construct a stop tree cache from a typed array
  *
- * @param {ArrayBuffer} data
+ * @param {Int32Array} data
  * @param {Number} size
+ * @returns {StopTreeCache}
  */
-export function create (data, size) {
-  const stopTreeCache = {}
-
-  stopTreeCache.data = data
-  stopTreeCache.index = new Int32Array(size)
+export function create ({
+  data,
+  size
+}: {
+  data: Int32Array,
+  size: number
+}): StopTreeCache {
+  const index = new Int32Array(size)
 
   // function that reads from array performing bounds checks
   const read = (idx) => {
@@ -26,7 +32,7 @@ export function create (data, size) {
 
   // build the index and de-delta-code
   for (let i = 0, pixel = 0, stopId = 0, time = 0; pixel < size; pixel++) {
-    stopTreeCache.index[pixel] = i
+    index[pixel] = i
     const nStops = read(i++)
 
     for (let stopIdx = 0; stopIdx < nStops; stopIdx++) {
@@ -37,5 +43,5 @@ export function create (data, size) {
     }
   }
 
-  return stopTreeCache
+  return {data, index}
 }
