@@ -1,123 +1,111 @@
-import jsolines, {getContour} from 'jsolines'
+// @flow
+import {getContour} from 'jsolines'
 import {createHandler} from 'web-worker-promise-interface'
 
 import accessibilityForGrid from './accessibility-for-grid'
-import getSurface from './get-surface'
-import getTransitiveData, {getPath, getPaths} from './get-transitive-data'
+import generateDestinationData from './generate-destination-data'
+import getIsochrone from './get-isochrone'
+import getPathsAndTimes from './get-paths-and-times'
+import {create as createSurface} from './surface'
 import isochroneTile from './isochrone-tile'
 import {create as createGrid} from './grid'
-import * as mercator from './mercator'
-import {create as createOrigin} from './origin'
+import {create as createOrigin, getStopPatternStopSets} from './origin'
 import {create as createStopTreeCache} from './stop-tree-cache'
+import {colorScheme} from './utils'
 
 module.exports = createHandler({
   putGrid (ctx, message) {
     if (!ctx.grids) ctx.grids = new Map()
-    ctx.grids.set(message.key, createGrid(message.grid))
+    const grid = createGrid(message.grid)
+    ctx.grids.set(message.id, grid)
+    const {contains, data, ...gridInfo} = grid
+    return gridInfo
   },
   accessibilityForGrid (ctx, message) {
     return accessibilityForGrid({
-      grid: ctx.grids.get(message.grid),
-      cutoff: message.cutoff,
-      surface: ctx.surface
+      ...ctx,
+      grid: ctx.grids.get(message.gridId),
+      cutoff: message.cutoff
     })
   },
   drawTile (ctx, message) {
     return isochroneTile(message.imageData, {
+      colorScheme,
       height: ctx.query.height,
       scaleFactor: message.scaleFactor,
-      surface: ctx.surface.surface,
+      surface: ctx.surface,
       width: ctx.query.width,
       xoffset: message.xoffset,
       yoffset: message.yoffset
     })
   },
-  setOrigin (ctx, message) {
-    ctx.origin = createOrigin(new Int32Array(message.arrayBuffer), message.point)
+  setOrigin (ctx, message, log) {
+    ctx.origin = createOrigin({
+      data: new Int32Array(message.data),
+      log,
+      point: message.point
+    })
+    const {data, index, ...originInfo} = ctx.origin
+    return originInfo
   },
   setQuery (ctx, message) {
     ctx.query = message.query
   },
   setStopTreeCache (ctx, message) {
-    ctx.stopTreeCache = createStopTreeCache(new Int32Array(message.arrayBuffer), ctx.query.width * ctx.query.height)
+    ctx.stopTreeCache = createStopTreeCache({
+      data: new Int32Array(message.data),
+      size: ctx.query.width * ctx.query.height
+    })
   },
   setTransitiveNetwork (ctx, message) {
     ctx.transitiveNetwork = message.network
   },
-  generateSurface (ctx, message) {
-    const { spectrogramData, ...surface } = getSurface({
-      grid: ctx.grids.get(message.grid),
-      origin: ctx.origin,
-      query: ctx.query,
-      stopTreeCache: ctx.stopTreeCache,
+  generateSurface (ctx, message, log) {
+    const {spectrogramData, ...data} = createSurface({
+      ...ctx,
+      grid: ctx.grids.get(message.gridId),
+      log,
       which: message.which
     })
+    ctx.surface = data.surface
+    ctx.waitTimes = data.waitTimes
+    ctx.walkTimes = data.walkTimes
+    ctx.inVehicleTravelTimes = data.inVehicleTravelTimes
 
-    ctx.surface = surface
-
-    return { spectrogramData }
+    return {spectrogramData}
   },
-  generateDestinationData (ctx, message) {
-    const { to, from } = message
-
-    const travelTime = ctx.surface.surface[to.y * ctx.query.width + to.x]
-    const waitTime = ctx.surface.waitTimes[to.y * ctx.query.width + to.x]
-    // NB separate walk time surface because some summary statistics don't preserve stat(wait) +
-    //   stat(walk) + stat(inVehicle) = stat(total)
-    const walkTime = ctx.surface.walkTimes[to.y * ctx.query.width + to.x]
-    const inVehicleTravelTime = ctx.surface.inVehicleTravelTimes[to.y * ctx.query.width + to.x]
-
-    return {
-      paths: getPaths({
-        origin: ctx.origin,
-        query: ctx.query,
-        stopTreeCache: ctx.stopTreeCache,
-        to
-      }),
-      transitive: getTransitiveData({
-        network: ctx.transitiveNetwork,
-        origin: ctx.origin,
-        from,
-        to,
-        query: ctx.query,
-        stopTreeCache: ctx.stopTreeCache
-      }),
-      travelTime,
-      waitTime,
-      inVehicleTravelTime,
-      walkTime
-    }
+  generateDestinationData (ctx, message, log) {
+    return generateDestinationData({
+      ...ctx, // origin, query, stopTreeCache, surface, transitiveNetwork
+      ...message, // from, to
+      log
+    })
   },
   getContour (ctx, message) {
     ctx.contour = getContour({
       cutoff: message.cutoff,
       height: ctx.query.height,
-      surface: ctx.surface.surface,
+      surface: ctx.surface,
       width: ctx.query.width
     })
     return ctx.contour
   },
   getIsochrone (ctx, message) {
-    return jsolines({
-      surface: ctx.surface.surface,
-      width: ctx.query.width,
-      height: ctx.query.height,
-      cutoff: message.cutoff,
-      interpolation: message.interpolation,
-      // coords are at zoom level of query
-      project: ([x, y]) => {
-        return [mercator.pixelToLon(x + ctx.query.west, ctx.query.zoom), mercator.pixelToLat(y + ctx.query.north, ctx.query.zoom)]
-      }
+    return getIsochrone({
+      ...ctx.query, // north, west, height, width
+      ...message, // cutoff, interpolation
+      surface: ctx.surface
     })
   },
-  getPath (ctx, message) {
-    return getPath({
+  getPath (ctx, message, log) {
+    return getStopPatternStopSets({
+      log,
       pathDescriptor: message.path,
       origin: ctx.origin
     })
   },
   getPaths (ctx, message) {
-    return getPaths({
+    return getPathsAndTimes({
       origin: ctx.origin,
       query: ctx.query,
       stopTreeCache: ctx.stopTreeCache,
